+++ examples/data-sources/harness_platform_connector_prometheus/data-source.tf
@@ -1,0 +1,3 @@
+data "harness_platform_connector_prometheus" "example" {
+  identifier = "identifier"
+}